+++ packages/@expo/cli/src/start/doctor/typescript/__tests__/TypeScriptProjectPrerequisite-test.ts
@@ -3,8 +3,6 @@ import { vol } from 'memfs';
 import * as Log from '../../../../log';
 import { TypeScriptProjectPrerequisite } from '../TypeScriptProjectPrerequisite';
 
-const asMock = (fn: any): jest.Mock => fn;
-
 jest.mock('../../../../log');
 
 describe('assertAsync', () => {
@@ -16,7 +14,6 @@ describe('assertAsync', () => {
   });
   it('skips setup due to environment variable', async () => {
     process.env.EXPO_NO_TYPESCRIPT_SETUP = '1';
-    asMock(Log.warn).mockClear();
     const prerequisite = new TypeScriptProjectPrerequisite('/');
 
     await prerequisite.assertAsync();